made overall structural and organizational changes to the code
Grouped important and major variables at the beginning. Everything is essentially grouped into the writePassword function.
Confirm prompts only execute once the password length is validated. A series of if statements follow to certify which conditions/parameters the user wants in the password. I used console logs to see what the results are at 2 different stages of the codes execution.  I learnt about the Join function and decided to use that function to put the generated password properly.